[AD-191] Integrate metadata persistence with DocumentDB (#80)
### Summary

[AD-191](https://bitquill.atlassian.net/browse/AD-191) Integrate metadata persistence with DocumentDB

### Description

With this change, schema is written to the same database as the one specified in the JDBC connection string. The schema is written to two collections `_sqlSchemas` and `_sqlTableSchemas`.

The connecting DocumentDB user will need the follow the built-in role `readWrite` (or greater) to be able to write the schema to the collections. If the user/driver is unable to write to the collections, the schema will be stored in memory.

The way to remove the schema is to drop the two collections:

```
use databaseName
db.getCollection("_sqlSchemas").drop()
db.getCollection("_sqlTableSchemas").drop()
```

### Related Issue

https://bitquill.atlassian.net/browse/AD-191